Combinediff now returns the "merged" diff
combinediff returns a new diff which represents the diff from the original
file and a new file which is the result of a merge of the changes from
the original diffs. Currenlty uses the "both" merge strategy.

Added in helper function createFromDiff which returns a file representation
of the diff (without any things which have been removed in the new file)