WIP: Add conflictAsPatch'
CAP' parallelizes conflicts based on PatchActions rather than
changehunks like CAP. This is necessary to fix the
prop_viewableConflict' case that failed where a delete file was filtered using
(>||< which currently uses the non-prime version as backend), that should have
been a conflict.

Previously findConflictsPA handled the case of dealing with tricky
PatchActions. That function has been removed, so it seems [unfortunately] that
some of the functionality needs to be done elsewhere, namely in
CAP'.

CAP' assumes it'll never see a CreateEmptyFile. Also, it assumes
it'll see at most 1 RemoveFile since if there were two, they shouldn't be in
conflict.  It handles the case of a single RemoveFile by turning it into a
ChangeHunk affecting the entire size of the file. The algoritm in
CAP' is largely the same as the one in CAP.

While prop_viewableConflict' (which uses CAP') compiles, it still
fails. At least now we can figure out that case too.
The types are super hairy though, as it can be seen in the prop code.