Fixes to conflictAsPatch
This is a different solution to the problems Hashem has shown with
getting a conflict as a patch. Instead of trying to use groupbypatch
and ungroupby, this uses the fact that because of the type class,
a conflict of parallel patches must all act on the same path. Therefore
we can strip out the path and pass it to the conflictAsPatch' function
which works on patchActions.

Small bug fix was made to conflictAsPatch' because it reversed the
conflict in an attempt to save lines, but this would swap how the
conflict is displayed to the user which would be confusing.

I don't really see the negatives of this solution except the property
of getConflict olds now can't really be checked unless some slightly
ugly removal of all the at path stuff. Now the code passes all
of our current quickCheck properties and compiles.