Change CreateEmptyFile to CreateFile
Needed cahnge because the patches were technically not in parallel because
we transformed adding a file into creating it empty and then the change
hunks.

Creates (and removes) conflict only when their contents are unequal.

Viewing the conflict between two CreateFiles is simply putting the first
one before the second to view because there is no common ancestor. This
could be a bit better, but will suffice for now.