Changed representation of WorkingTree
FileSystem now includes a set of tracked paths and loads in
those files being tracked on startup.

Added in serialization and saving extras.

In order for this to make sense, need to add a new command called "nor add"
to stage changes before being allowed to commit them. Currently that is now
the only way to commit.