Updating Test and Cases
Test and Cases now work with dataCommit

Spent a long time trying to find bugs in how I was testing.

Turns out that because of how the patches are regenerated, just
because they were created without conflicts doesn't mean that
there won't be conflicts laters as rebase progresses because
applying other patches can change the diff algorithm's manner
for creating change hunks. The rebase appears to work if that
never happens, but that is very difficult to know before hand.